+++ Node.java
@@ -1,20 +1,20 @@
 /**
  * Node.java
  *
- * CS4341: Project 1
+ * CS4341: Project 2
  * Group: Zachary Robbins, Kyle McCormick, Elijah Gonzalez, Peter Raspe
  */
 
 public class Node {
 
-	public final int value;
+	public final double value;
 	public final Operation operation; // NULL if this is root
 	public final Node parent;	      // Parent node, NULL if this is root
 	public final int depth;
 	public Node[] children;           // Children, determined by operators list from parsing.
 	                                  //  NULL if not expanded
 
-	public Node(int value, Operation operation, Node parent) {
+	public Node(double value, Operation operation, Node parent) {
 		this.value = value;
 		this.operation = operation;
 		this.parent = parent;
