+++ Search.java
@@ -1,18 +1,18 @@
 /**
  * Search.java
  *
- * CS4341: Project 1
+ * CS4341: Project 2
  * Group: Zachary Robbins, Kyle McCormick, Elijah Gonzalez, Peter Raspe
  */
 
 public abstract class Search {
 
-	int startValue;
-	int targetValue;
+	double startValue;
+	double targetValue;
 	double timeLimit;
 	Operation[] operations;
 
-	public Search(int startValue, int targetValue, double timeLimit, Operation[] operations) {
+	public Search(double startValue, double targetValue, double timeLimit, Operation[] operations) {
 		this.startValue = startValue;
 		this.targetValue = targetValue;
 		this.timeLimit = timeLimit;
